UPDATE
Added auto Installer function

+++ Shypy/packages/utils/utils.py
@@ -2,6 +2,7 @@ import os
 import sys
 import time
 import datetime
+import importlib
 from itertools import cycle
 
 from ..variables.variables import *
@@ -32,6 +33,25 @@ class Utils:
             self.write(message="Exited from Shypy", level=2, clear=True)
             sys.exit(1)
             
+    def module_exist(self, module : str) -> bool:
+        """
+            Check if a module exists.
+
+            This inner function checks if a given module exists.
+
+            Args:
+                module (str): The name of the module to check.
+
+            Returns:
+                bool: True if the module exists, False otherwise.
+            """
+        try:
+            importlib.import_module(module)
+            return True
+        
+        except ImportError:
+            return False
+    
     def write(self, message : str, level : int, delay : float = 0, clear : bool = False) -> None:
         """
         Prints a message with a specified level.
@@ -77,6 +97,19 @@ class Utils:
                 
         time.sleep(delay)
         
+    def auto_install(self, libraries : list[str]) -> None:
+        """
+        Installs missing requirements.
+
+        Args:
+            libraries (list[str]): Missing libraries list.
+            
+        Returns:
+            None
+        """
+        self.write(message="Loading requirements please be patient", level=1)
+        for lib in libraries:
+            os.system(f"pip install {lib}")
     
     def check_modules(self) -> None:
         """
@@ -117,36 +150,36 @@ class Utils:
             except (KeyboardInterrupt):
                 self.write(message="Exited from Shypy", level=2, clear=True)
                 sys.exit(1)
-        
-        def exist(module: str) -> bool:
-            """
-            Check if a module exists.
-
-            This inner function checks if a given module exists.
-
-            Args:
-                module (str): The name of the module to check.
-
-            Returns:
-                bool: True if the module exists, False otherwise.
-            """
-            try:
-                __import__(module)
-                return True
-            
-            except ImportError:
-                return False
 
         try:
             loader()
             self.clear()
-            missing_libs : str = [module for module in LIBS if not exist(module)]
-
+            missing_libs : list[str] = [module for module in list(LIBS.keys()) if not self.module_exist(module)]
+            missing_list : list[str] = [LIBS[lib] for lib in missing_libs] 
+                
             if missing_libs:
-                self.write(message=f"The following libraries are missing: {', '.join(missing_libs)}", level=4)
-                self.write(message=f"Please run {COLOR['YELLOW']}'pip install -r requirements.txt'{COLOR['RESET']} or {COLOR['YELLOW']}'pip install {', '.join(missing_libs)}'{COLOR['RESET']}", level=3)
-                self.write(message="You need to install the required libraries to continue.", level=5)
-                sys.exit(1)
+                self.write(message=f"The following libraries are missing: {', '.join(missing_list)}", level=4)
+                self.write(message=f"Please run {COLOR['YELLOW']}'pip install -r requirements.txt'{COLOR['RESET']} or {COLOR['YELLOW']}'pip install {', '.join(missing_list)}'{COLOR['RESET']}", level=2)
+                try:
+                    self.write(message="Auto-install the requirements [y/n]\n", level=3)
+                    choice = input(f"{COLOR["CYAN"]}Shypy >> {COLOR["RESET"]}").lower().strip()
+                    
+                    if choice == "y":
+                        self.clear()
+                        self.auto_install(libraries=missing_list)
+                        
+                    else:
+                        self.clear()
+                        self.write(message="You need to install the required libraries to continue.", level=5)
+                        sys.exit(1)
+                
+                except EOFError:
+                    self.write(message="Exited from Shypy", level=2, clear=True)
+                    sys.exit(1)
+                
+                except KeyboardInterrupt:
+                    self.write(message="Exited from Shypy", level=2, clear=True)
+                    sys.exit(1)
                 
             else:
                 self.write(message="All required requirements are already installed.", level=1)
@@ -154,3 +187,4 @@ class Utils:
 
         except KeyboardInterrupt:
             self.write(message="Exited from Shypy", level=2, clear=True)
+            sys.exit(1)
